backend/requirements.txt
fastapi
uvicorn[standard]
aiomysql
PyMySQL
cryptography
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
pydantic
pydantic-settings
httpx
twilio
sendgrid
apscheduler
